Day 9 uses negative X and Y values, so let's account for that in Position. The fact that every other solution only uses non-negative numbers shouldn't be affected by this... right?

+++ 2018/AdventOfCode/AoCHelpers/Position.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "Position.h"
 
-advent_of_code::Position::Position(const unsigned int x, const unsigned int y) : m_x{ x }, m_y{ y }
+advent_of_code::Position::Position(const long long x, const long long y) : m_x{ x }, m_y{ y }
 {}
 
 bool advent_of_code::Position::operator<(const Position& other) const {
@@ -25,11 +25,11 @@ bool advent_of_code::Position::operator!=(const Position& other) const {
 }
 
 bool advent_of_code::Position::is_adjacent_to(const Position& other) const {
-   if (m_x == other.m_x && 1 == abs(static_cast<int>(m_y) - static_cast<int>(other.m_y))) {
+   if (m_x == other.m_x && 1 == abs(m_y - other.m_y)) {
       return true;
    }
 
-   if (m_y == other.m_y && 1 == abs(static_cast<int>(m_x) - static_cast<int>(other.m_x))) {
+   if (m_y == other.m_y && 1 == abs(m_x - other.m_x)) {
       return true;
    }
 
@@ -46,8 +46,7 @@ std::vector<advent_of_code::Position> advent_of_code::Position::get_adjacent_pos
 }
 
 size_t advent_of_code::Position::distance_to(const Position& other) const {
-   return (abs(static_cast<int>(m_x) - static_cast<int>(other.m_x))
-         + abs(static_cast<int>(m_y) - static_cast<int>(other.m_y)));
+   return static_cast<size_t>(abs(m_x - other.m_x) + abs(m_y - other.m_y));
 }
 
 std::wostream& advent_of_code::operator<<(std::wostream& out, const advent_of_code::Position& p) {
