+++ 2018/AdventOfCode/AoCHelpers/Position.h
@@ -8,7 +8,7 @@ namespace advent_of_code {
 struct Position
 {
    Position() = default;
-   Position(const unsigned int x, const unsigned int y);
+   Position(const long long x, const long long y);
    Position(const Position& p) = default;
    bool operator<(const Position& other) const;
    bool operator==(const Position& other) const;
@@ -20,8 +20,8 @@ struct Position
 
    friend std::wostream& operator<<(std::wostream& out, const Position& t);
 
-   unsigned int m_x;
-   unsigned int m_y;
+   long long m_x;
+   long long m_y;
 };
 
 #ifdef max
@@ -29,5 +29,5 @@ struct Position
 #endif
 
 const Position NONE_POSITION{
-    std::numeric_limits<unsigned int>::max(), std::numeric_limits<unsigned int>::max() };
+    std::numeric_limits<long long>::max(), std::numeric_limits<long long>::max() };
 }
